+++ test/src/solid_ui_object.dart
@@ -1,0 +1,11 @@
+//Copyright (C) 2012 Sergey Akopkokhyants. All Rights Reserved.
+//Author: akserg
+
+part of dart_web_toolkit_test;
+
+/**
+ * Clear UiObject class implementation.
+ */
+class SolidUiObject extends ui.UiObject {
+  
+}